pyproject.toml
[tool.poetry]
name = "newWebApp"
version = "0.0.1"
description = "Youtube downloader able to download intervals"
authors = ["Christian Jørgensen <friis93@me.com>"]

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1.1"
pafy = "^0.5.5"
youtube-dl = "^2020.1.24"
ffmpeg-python = "^0.2.0"
gunicorn = "^20.0.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
